Implement correct job cancellation flow for non-suspendable Jet jobs [HZ-3679] [HZ-3882] (#26037)
This patch introduces a correct non-suspendable job cancellation control
flow by terminating job if `CancellationByUserException` were caught. It
would be better to allow `SqlResult` to join the job, but the jobId is
not available for `SqlResult`. It is planned to be fixed in the next
release, TDD [is
available.](https://github.com/hazelcast/hazelcast/pull/25777)

It also implements `SqlSummary` for normal Jet jobs if they are created
from SQL: it affects both `CREATE JOB` and `ANALYZE`
---------
Signed-off-by: Sasha Syrotenko <oleksandr.syrotenko@hazelcast.com>
Co-authored-by: Krzysztof Jamróz <krzysztof.jamroz@hazelcast.com>