CompositeMessageTaskFactory unwanted litter.
The factories hashtable is of type Int2ObjectHashMap; which is great so
that no litter is created based on the integer key.

But the field is of type Map and Map doesn't have a Map.get(int). So there
is still boxing going on which causes litter.

This PR fixes that by changing the type of the factories field to
Int2ObjectHashMap.