Litter reduction in CompositeMessageTaskFactory.create (#26129)
The factories hashtable is of type Int2ObjectHashMap; which is great
because no litter is created based on the 'int' key.

But the field is of type Map and Map doesn't have a Map.get(int) only
Map.get(Object). So there is still boxing going on which causes litter.
So effectively the Int2ObjectHashMap is in vain.

This PR fixes that by changing the type of the factories field to
Int2ObjectHashMap.