[TPC] Fixing performance regression [HZ-3909] (#181)
Import of https://github.com/hazelcast/hazelcast/pull/26033

**Original PR description:**

Fixes TPC performance regression compared to the tpc-engine-advanced
branch.

There are 2 fixes:
1) Allow for passing any object to the AsyncSocket instead of requiring
IOBuffer. THis will reduce the amount of litter/data copying. For this
purpose the AsyncSocketWriter was introduced so that an arbitrary object
can be passed to the sendBuffer. It will also reduce TCP overhead on
small packets because instead of writing an array of buffers, they are
first merged into the write buffer.

2) TPCOperationScheduler didn't do any batching (batch size is 1). Now
batching is done based on a time period of 500us.

For enterprise PR see:
https://github.com/hazelcast/hazelcast-enterprise/pull/6906

Imported changes:

- 1f651e0b5615a30fc34a5e7a49297ae6e00bb025 [TPC] Fixing performance
regression [HZ-3909]

Closes hazelcast/hazelcast#26033

[HZ-3909]:
https://hazelcast.atlassian.net/browse/HZ-3909?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Peter Veentjer <alarmnummer@gmail.com>
Co-authored-by: Jack <jack.green@hazelcast.com>
GitOrigin-RevId: 25b070d6a56da0b3d312eb498f9fbc07f0f445f9