[TPC] Fix NPE on onFinishConnect. (#244)
The problem was that with the NioAsyncSocket onFinishConnect, the
OP_CONNECT was not removed from the interest-set since we just added the
OP_READ to the interest-set.

This has been changed: we only care for OP_READ and need to remove
anything else.

This unwanted remaining registration of the OP_CONNECT leads to an
OP_CONNECT ready event when the socket closes. And this is when it will
run into an NPE because connectFuture is null (since that field is
nulled after the connect).

The NPE only leads to noise in the log when closing a socket. Has no
other consequences.

Fixes https://github.com/hazelcast/hazelcast-mono/issues/243

GitOrigin-RevId: 759e67be38b2fbc13d8827289588ca87abcdc9d7