Explicit External IP check in Kubernetes Discovery Plugin [CN-296] (#211)
Import of https://github.com/hazelcast/hazelcast/pull/25324

**Original PR description:**

The problem is that it assumes the service has a type of LoadBalancer
first. If it cannot find ExternalIP there, it jumps into catch block,
assumes it has a NodePort kind, and determines ExternalIP according to
it.

The block needs to be improved in such way:
It needs to check the kind of the service, and use a specific method to
determine External IP.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Label `Add to Release Notes` or `Not Release Notes content` set
- [x] Request reviewers if possible
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases

Imported changes:

- f63f7fad73487ddbc66f0b31d8c1b55f17999597 Explicit External IP check in
Kubernetes Discovery Plugin...

Closes hazelcast/hazelcast#25324

Co-authored-by: semihbkgr <buyukgungorsemih@gmail.com>
GitOrigin-RevId: 2bb2877cfb1f876e9263616b64307ae3a1a386f8