Remove timeout limit from `MavenInterface` (#271)
`MavenInterface` is used to download additional dependencies for test
execution, but the downloads have been failing due to an internal two
minute timeout on the download process.

During test execution, it's being accessed via
`HazelcastVersionLocator.locateVersion(String, boolean)` which is
downloading ~28MB of data.

Due to the other load on the environment while tests are running, it's
not possible to guarantee that the download will complete within this
window, and Maven already has a timeout mechanism that is aware of the
state of the connection, so the two minute timeout is redundant.

Moreover, if the operation **can** take more than two minutes, any tests
relying on it probably shouldn't be `QuickTest`s.

Changes:
- Remove two minute timeout
- Make any dependent tests `NightlyTest`s

Fixes: https://github.com/hazelcast/hazelcast/issues/25919
GitOrigin-RevId: 04ab41e1724b222c3f07a5fc4165bcd5a3f3c662