Update leader stickiness algorithm [HZ-4083] (#326)
An existing leader failure detection algorithm and subsequent
`PreVoteTask`/`VoteTask` work in the following way.

`LeaderFailureDetectionTask` is a periodic task, running every 2-3 sec
(_leader-election-timeout-in-millis + random_) on all Followers:
- task checks that the leader is reachable via
`raftIntegration.isReachable(leader)`
- task checks that `lastAppendEntriesTimestamp +
maxMissedLeaderHeartbeatCount * heartbeatPeriodInMillis <
currentTimeMillis`
If the check fails it runs `PreVoteTask`.

The `PreVoteTask` has protection from disruptive followers - the
so-called _leader stickiness_ check. This check should only give a vote
for the candidate if the current voting follower also agrees that the
leader is not available.

The current implementation of the check looks like this:
`if (raftNode.lastAppendEntriesTimestamp() > Clock.currentTimeMillis() -
raftNode.getLeaderElectionTimeoutInMillis()) {...}`

This check is not valid, because `getLeaderElectionTimeoutInMillis()`
returns a random number from 2 to 3 seconds, but the
`lastAppendEntriesTimestamp` (without additional activity) is updated
every 5 seconds. As a result, the follower can give a vote for the
candidate even if the leader is alive.

To fix, this the leader stickiness should use the same checks as a
`LeaderFailureDetectionTask`:
- checks that the leader is reachable via
`raftIntegration.isReachable(leader)`
- checks that `lastAppendEntriesTimestamp +
maxMissedLeaderHeartbeatCount * heartbeatPeriodInMillis <
currentTimeMillis`

GitOrigin-RevId: aa73ff7ed601f9b80b27c08a6cab8e4748fd2ea0