Add support for URL-only client->member NS resources (#283)
This commit makes adjustments to `ResourceDefinition` classes which
allows for `payload` to be `null`, and has the member read the payload
from the provided member-relative path instead.

This is achieved by casting a client-side `UserCodeNamespacesConfig` to
its dynamic variant, `ClientDynamicClusterUserCodeNamespacesConfig` and
calling `addMemberRelativeNamespaceConfig` with
`ResourceDefinitionHolder` objects that contain a `null` payload and
`non-null` `url`. The casting is involved because we generally don't
want normal public APIs to be able to use this approach, due to
potential for confusion (as paths must be member-relative).

Fixes https://hazelcast.atlassian.net/browse/HZ-4022

GitOrigin-RevId: 5e9def735d62f97af82a6eab4dbc9e9fef3b68eb