Do not put any member without CLIENT endpoint qualifier configured into the member list [API-2094] (#364)
Do not put any member without CLIENT endpoint qualifier configured into
the member list published by the cluster view listener. The client will
only receive and know the members with which it can establish connection
to. The partition table may contain the member without client endpoint,
if client needs to send a request to that partition then it will be sent
to a random member since there will be no direct connection available to
that member. The test verifies that it works.

fixes https://github.com/hazelcast/hazelcast/issues/25032

GitOrigin-RevId: e4161b3ca9beeee6f0bf6e08c246326cd6181c0d