Collect job execution metrics from running participants only [HZ-3984] (#194)
**Problem:**
A job is initiated for all participants, and these participants may
process the job at different speeds, resulting in different completion
times. When a metrics request is made to gather job data from all
participants, an error response from one participant may indicate that
the job for that participant has either already been completed or not
started. In such cases, the master attempts to re-query all participants
until successful responses are received from everyone or until the
overall job status changes to complete. However, if the job is completed
on one participant and running endlessly (streaming) on another, the
metrics retrieval process enters an infinite loop.

**Solution:**
Metrics are stored in the context once a participant completes the job.
When requesting metrics, the master only queries participants whose jobs
have not been completed; for completed jobs, metrics are retrieved from
the context. This resolves the issue where one participant has finished
the job while another is running endlessly.

In cases where the job on a participant has already been completed, but
this information has not reached the master, the master handles it by
reattempting the request to the participant. The number of attempts is
limited by the total number of participants, ensuring sufficient
attempts to retrieve the data.

In the event that, despite all attempts, we still encounter an error, it
indicates a critical process error and returns an error in this case.

Fixes hazelcast/hazelcast#26126 [[HZ-3984]]

**Checklist:**
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases

[HZ-3984]:
https://hazelcast.atlassian.net/browse/HZ-3984?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: 0fdec45c6d8c50572309d04e34f14bf2ec203038