Support Eclipse IDE in SQL module - prevent compilation issue (#350)
When `hazelcast-sql` is opened in Eclipse, the following compilation
errors prevent execution:
- `HazelcastSqlParser` cannot be resolved to a variable
- The import `com.hazelcast.jet.sql.impl.calcite` cannot be resolved

`HazelcastSqlParser` is a class generated during the Maven build and
ends up in `target/generated-sources/java`, but Eclipse does not
understand this directory needs to be on the classpath to resolve it.
The classpath used in Eclipse is dynamically generated by it's Maven
plugin.

To address this, `build-helper-maven-plugin` is used to describe the
configuration required.

---------

Co-authored-by: Josef Cacek <josef.cacek@gmail.com>
GitOrigin-RevId: 31635a4db6f9e0dfd2e644a14bb46e8d341ef032