Turn off deprecated warnings (#406)
There are many warnings in the build due to deprecated things in our
code base (e.g. old user code deployment). This is just noise.

This will also turn off warnings where the deprecated item comes from
either the JDK or 3rd party library, but historically we only fix these
items when the items are actually removed, so it seems to be a good
trade-off.

GitOrigin-RevId: 21c47a57d336ad3a8f2f73344b64e9f6ba9c0c80