Remove support for removing Namespaces at runtime [HZ-4176] (#398)
After further discussion, the UCN team has agreed that there is not a
real use-case for removal of a Namespace at runtime. The fact that data
structure configs cannot de-reference a Namespace once referenced, and
that the general design pattern in HZ is to support "add" or "replace"
functionality but not remove, it shouldn't be an expectation anyway.

If we were to keep this functionality it would require additional work
to support features such as Hot Restart, and we don't believe this is a
worthwhile effort when there is no use-case for the feature.

GitOrigin-RevId: 93b74b0cacf49d0d3273393e2de5b81f570ffc7b