Create a fallback option for `CREATE MAPPING` with portable format [HZ-4026] (#397)
This patch fixes an issue with the impossibility of resolving `Portable`
fields, if columns were not explicitly defined during mapping creation
**AND** a containing map was not used yet (no entry in
`PortableContext`). Speaking frankly, `Portable` class registration
system is lazy by design, but in the SQL context we want to eagerly
create class definitions.

The problem was solved with a dirty trick: during the fields resolving
process, if we were not able to detect the class definition, we manually
create the `Portable` object and use it to forcefully register the class
definition with the created object (via
`lookupOrRegisterClassDefinition`). Then, the object is thrown
immediately.

Still, there is a use case (and, it is pretty common), when default
constructor doesn't initialize fields with non-null values in `Portable`
object. Here, we're done and we interrupt execution with clear
exception.

GitOrigin-RevId: c0a1ac72c06e0c192c17e06d7bf9fafd51929b1d