Remove unnecessary `maven-enforcer-plugin` Java 8 configuration (#469)
In https://github.com/hazelcast/hazelcast/pull/23400,
`maven-enforcer-plugin` was configured to enforce that all of the
projects' bytecode - including dependencies - was compatible with Java
8, to ensure compatibility with a Java 8 runtime.

Some dependencies were multi-release and include Java 9 bytecode, but
were still compatible with Java 8 - as such an exclusion was added, to
avoid these failing the build.

Now the application is Java 11, these dependencies are no longer an
issue and as such do not need to be ignored.

The only exclusion that needs to remain is `org.bson.codecs.record.*`,
which requires Java 17.

`maven-enforcer-plugin` is bound to the `validate` phase so is covered
by the PR builder and when the JAR is packaged.

Originally raised as [PR
feedback](https://github.com/hazelcast/hazelcast-mono/pull/436#pullrequestreview-1829105228).

GitOrigin-RevId: c155b69c86d89137f268733beec53df20890860e