Hot Restart namespaces persistence [HZ-3680] (#402)
HZ-3680 Add persistence of Namespaces for Hot Restart

Data structures eligible for Hot Restart Persistence (Cache & IMap) have
their configs stored on disk and reloaded in the case of a cluster
restart.

User Code Namespaces (UCN) introduced specific namespace referencing on
a DS. In addition to static configuration, Namespaces can be updated or
added dynamically. Therefore there a certain cases in which namespaces
must be persisted as part of a map/cache HR config persistence.

Examples of such cases :

(In each case a map or cache would reference the namespace and be Hot
Restart enabled)

1/ A statically added Namespace which is referenced by a HR enabled DS
has been updated dynamically.
2/ A dynamically added Namespace has been been referenced b a
dynamically added DS with HR enabled.

NOTE: when a dynamically added HR data structure is added which
references a statically configured namespace, it is not necessary to
persist the Namespace. Therefore, we consider only those namespaces
which have been updated/added by dynamic configuration to be candidates
for persistence.

This commit enables the above functionality by extending the
capabilities of `DynamicConfigListener` to cater for Namespace
persistence when handling `IMap` or `ICache` config persistence. In
addition, Namespaces can be added or overwritten through this listener.
`PersistentConfigDescriptors` is extended to handle the case of
overwriting config files.

On persistence reload, it is necessary to order the notification to
listeners to ensure the dependency chain is preserved. Otherwise an
error could occur where a HR enabled DS can not find a Namespace on the
system.

In addition to testing the new overwrite functionality of
`PersistentConfigDescriptors`, unit test coverage has been added for all
other public methods which has previously been missing.

The `UserCodeNamespaceService::addNamespaceConfig` has been updated to
add the `UserCodeNamespaceConfig` to the `UserCodeNamespacesConfig`. A
new protected method `addNamespaceRaw` was added to the
`UserCodeNamespacesConfig` so that we could bypass broadcasting this to
all other members, as it is an unnecessary step.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: 1fe589dd11a6c41d42702588910932f04b6f6f2a