Integrate namespaces with dynamic configuration persistence [HZ-4028] (#437)
Adds support for dynamic persistence of namespaces if configuration
persistence is enabled.
Fixes issues in yaml generation where namespaces name was set to lower
case.
Removes support for dynamically adding namespace if namespaces is
disabled.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible

---------

Co-authored-by: Jack <jack.green@hazelcast.com>
GitOrigin-RevId: ae337b155d676fa5c5f3acf07e437201585fdbdc