Avoid boxing of partitionId in getPartitionAssignment when partition pruning is not used (#486)
`getPartitionAssignment` is executed each time a light job is started
and with default partition count (271) more than half of the values is
not cached. For simple SQL queries which use jet jobs it account for
approximately 5% of all allocations.

For the case without partition pruning this PR restores previous way of
iteration, before introducing partition pruning.

Allocations from JFR before change:

![image](https://github.com/hazelcast/hazelcast-mono/assets/79092062/4caa433f-911a-4afe-970a-86148d7e9e4c)

After change `Integer` is gone from the top of the stats.

The throughput is not significantly improved (`once-worker` is a
baseline), and the latencies are slightly better (note however that this
was a throughput test not latency test), possibly due to less work for
GC.

run_label | once-worker (baseline) | no-boxing
-- | -- | --
benchmark | sqlitetable.select | sqlitetable.select
10%(us) | 10,633.22 | 10,387.45
20%(us) | 10,911.74 | 10,674.17
50%(us) | 11,460.61 | 11,255.81
75%(us) | 11,845.63 | 11,649.02
90%(us) | 12,263.42 | 12,050.43
95%(us) | 12,771.33 | 12,500.99
99%(us) | 15,728.64 | 15,327.23
99.9%(us) | 19,660.80 | 19,775.49
99.99%(us) | 25,788.42 | 25,214.97
max(us) | 34,439.17 | 30,097.41
operations | 3,055,756 | 3,148,445
duration(ms) | 96007 | 97007
throughput | 31,828.47 | 32,455.85

![image](https://github.com/hazelcast/hazelcast-mono/assets/79092062/7bff3aab-1c65-4e8f-91f1-4e7d7cb25159)

Surprisingly, this does not show for light jobs uploaded from clients.
Maybe there are other overheads (serialization, flake id) or maybe TLAB
is used, which was not enabled in the JFR profile.

GitOrigin-RevId: 8faa91acd8f6ee06fb7dc2b55f493915d13c42cb