ReplicatedMap fetchEntryViews API [HZ-3932] (#188)
Adds ClientReplicatedMapProxy.fetchEntryViews. This is an internal API to fetch entryviews (entry data + metadata) needed for data migration tool. It's merged into master because we want to make this available in the future so that we won't require a jar that customers use to restart their cluster with.

See client protocol pr:
https://github.com/hazelcast/hazelcast-client-protocol/pull/492

GitOrigin-RevId: 50947779d2b09bd27257f1aa1ea4a9deba3abba9