HD and TS licensing changes for 5.4 [HZ-4038, HZ-4039] (#509)
This patch implements an updated licensing for 5.4. It includes changes
for memory consumption licensing of HD memory and Tiered Storage. The
solution is to launch scheduled checks, which happen once per minute for
each feature. We log 80%, 90%, and 100% limits of memory consumption
with different log levels and messages (they are described in
requirements).

**HD**

Requirements:
- [x] HD capacity licensing is calculated at the member level.
- [x] HD capacity usage includes only the native memory pool, limited by
200 Gigs per member.
- [x] Users should be notified in advance that they are nearing the
limit of their license. 
- [x] License check is performed at start up and once per minute.
- [x] For 5.4, we will not be enforcing the license. In the event of a
license breach, we will only issue a warning

**Tiered Storage**

According to the requirement to use per-cluster charging, the author
opted for the usage of service `ReplicatedMap<Uuid, Long>`. In this
structure, the key corresponds to the member's `UUID`, while the value
denotes the current memory consumption in bytes, encompassing Hybrid Log
and Device storage. Each member contributes to this map at minute
intervals, and the master member undertakes the computation and license
validation. Additionally, the master member takes the responsibility of
monitoring member vitality and purging memory consumption records for
departed cluster members. This presents a straightforward and cheap
solution, with the payload size of the `ReplicatedMap` in kilobytes even
for large clusters — essentially negligible.

Worth noting, that following the specified requirements, the license
check must be paused in the event of a split-brain scenario within the
cluster. While Hazelcast member lacks real-time awareness of its
split-brain status, the proposed solution seamlessly aligns with
split-brain occurrences, as each temporary master erases map records of
members in other 'subcluster'.

Requirements:
- [x] TS capacity licensing is calculated at the cluster level, ie. an
aggregation of the used capacity across all configured devices in the
cluster. 
- [x] TS capacity usage includes inflight memory (Hybrid Log) and data
written to disk (Device). 
- [x] For now, we will simply use the directory size for device usage
calculations, which includes data, garage, indexes and metadata.
- [x] Users should be notified in advance that they are nearing the
limit of their license. 
- [x] License check is performed at start up and once per minute.
- [x] Prevent the start up of the cluster if the memory tier is not <=
disk tier.
- [x] For 5.4, we will not be enforcing the license. In the event of a
license breach we will only issue a warning:
- [ ] ~~License calculations should only be performed while the cluster
is healthy, so that problems like a split brain do not produce temporal
out of license events.~~
GitOrigin-RevId: 20ffacdb34daea90f0c21b1a095ded3659f39d6b