Add getCPGroupIds to CPSubsystem [HZ-3937] (#591)
This exposes already existing API
`CPSubsystemManagementService#getCPGroupIds` in CPSubsystem, where it is
easier to use and is supported both by the member and client.

This API provides access to all active group ids, which can then be used
to list all data structures within the group. This API will be sent in a
separate PR.

Protocol API:
https://github.com/hazelcast/hazelcast-client-protocol/pull/504

GitOrigin-RevId: c3d7eb5a55764b834bf96562d782424297932563