Consolidate checkstyle configuration (#574)
Hazelcast OS & EE both use checkstyle, but with their own, slightly
divergent, configurations.

Centralised to use the OS checkstyle configuration file, with a small
tweak to the header configuration to support the difference between
formats of OS & EE header requirements.

No changes were required to the OS `checkstyle.xml`, but a handful of
elements had to be moved from the (now deleted) EE `suppressions.xml` to
the OS variant.

---------

Co-authored-by: Tomasz Gawęda <tomasz.gaweda@outlook.com>
GitOrigin-RevId: 33e6eafc3b6c76510e9dc48292f4cae3cfe88934