CancellationByUserException suppresses in logs [HZ-3595] [HZ-1994]  (#661)
Fixes: HZ-3595, HZ-1994

CancellationByUserException is set as a technical exception, so it is
not logged after the operation is canceled. Affects both normal and
light jobs. Looks like there is enough information in the logs about
canceled operations there is no need for this
CancellationByUserException in the log.

Additionally, if the job was canceled then logged that it was canceled
instead of failed.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases
GitOrigin-RevId: e29364d220f9328ad82d54d1adf1f1e2cad30b17