Execute SqlJDBCDataConnectionTest on arm64 for supported databases (#662)
Database containers are created eagerly by `SqlJDBCDataConnectionTest`
and MSSQL is not supported. However, because it generated assumption
failure during parameters generation, entire test was skipped.

`DatabaseRule` was made lazy as it is now used in `parameters()` method.
`@Parameterized.Parameters` method is invoked before `@Category`
filtering - in all builds.

Additionally, on MacOS I observed race conditions when Postgres
container was started - port on the host was available later than the
database reported that it is ready. This resulted in rejected
connections and test failure. To fix that Postgres provider waits both
for port availability and log message.

GitOrigin-RevId: ff8a8906e07af87f13d46cd0e836a84361b6e1dc