Add ClientReplicatedMapProxy.putAllWithMetadata [HZ-4294] [5.3.z] (#644)
Cherry-picks
https://github.com/hazelcast/hazelcast-mono/commit/b063928bd6d86959ec66ab222cfa072e63a656f4
except client protocol compatibility test related changes:
referenceobjects, binary files, compatilibity test files. We don't
change them because adding new API in patch is not possible normally
because we don't have maintenance branches in client protocol repo.

We need this API for data migration tool and it's an internal API. So we
will be ok if we ignore the compat. tests for this API.

---------

Co-authored-by: devOpsHazelcast <devops@hazelcast.com>
GitOrigin-RevId: 56214a9fbf44f0a42a5d6faf7891147e9e9cf010