Provide metrics for namespaces [HZ-4023] (#625)
This PR provides metrics for namespace configurations.

Specifically the following information is provided :

Namespace **updateTime** (key = updateTime)
Namespace **resourceCount** (key = resourceCount)
Listed `ResourceStats` for a namespace.

`ResourceStats` provides :
Size of resource in bytes.
Type (ordinal) of resource.

On the `MemberState` a list of existing namespace ids will be published.

The statistics are captured by default and updates to an existing
namespace via dynamic configuration will overwrite the previously
recorded metrics for the specific Namespace.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: 9afdcb3e99bd40afc50e9b62eba6af244dbc44ab