Support WAN ref updates in UpdateMapConfigOperation [HZ-4221] (#544)
There was a need for Hazelcast Operator to be able to update the
`WanReplicationRef` of an existing `IMap` via its `MapConfig` - this
commit adds support for this within the `UpdateMapConfigOperation`,
provided through a separate construction to minimize negative impact on
existing Management Centre integration.

The `applyWanReplicationRef` variable exists as part of our RU_COMPAT
requirement, as a `null` `WanReplicationRef` is valid (meaning to remove
the reference), so to ensure we do not erroneously remove a WAN
replication reference, this field exists to decide whether we apply the
reference changes or not.

The related `UpdateMapConfigMessageTask` and associated codec have both
been updated to support functionality.

A simple regression test has also been introduced in the
`DynamicMapConfigTest`.

Client protocol PR:
https://github.com/hazelcast/hazelcast-client-protocol/pull/508
Closes https://hazelcast.atlassian.net/browse/HZ-4221

GitOrigin-RevId: 780512d33db47a576d1c604b6f9665e1be5f9f5e