Fix `XmlYamlConfigBuilderEqualsTest` URLs on Windows [HZ-4255] (#663)
In https://github.com/hazelcast/hazelcast-mono/pull/330, the URLs used
to specify the location of a resource to load into a user code namespace
were converted from being statically declared inside the configuration,
to be dynamically generated to support dynamic compilation/generation of
these resources.

Specifically, for a given resource, the expected URL in the
configuration file (some relative `file:///folder/file.jar`), and the
actual location the resource would be found at (some absolute
`file:///opt/temp/folder/file.jar`) were generated dynamically, using
`Path.toURI()`.

This worked on all platforms, but Windows - it's implementation of
`Path.toUri()` is **unable** to create a URI to a relative path (i.e.
the first thing it does is convert the path to an absolute path), and
any attempt to trick it using `relativize` etc will cause an exception
if the parent/child directories are of different types (e.g. relative,
absolute, UNC etc).

To resolve this, the expected URL in the configuration file is now
generated via String concatenation, with a comment to explain the above.
This change affects only test code.

In addition, the `OS.ensureUnixSeparators(String)` method - that
replaces Windows separators with Unix ones - was refactored to avoid
hardcoding the Windows separator as the Java filesystem framework
already knows all of this. Functionally equivalent, but if it _did_
change in future, we'd be covered.

Fixes [HZ-4255](https://hazelcast.atlassian.net/browse/HZ-4255),
https://github.com/hazelcast/hazelcast/issues/26214

[HZ-4255]:
https://hazelcast.atlassian.net/browse/HZ-4255?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: 79ef8c02556917385399d50fe0d0271b9b89d9d9