Make KeyedWindowResult hashCode conform to Map.Entry contract [HZ-4335] (#697)
`KeyedWindowResult` implements `Map.Entry` but this seems to be done
only for convenience. This is a public API so users can use this as
well.

[Upgraded](https://github.com/hazelcast/hazelcast-mono/pull/382)
`EqualsVerifier` (which is needed for Java 21) marks non-conformance
which cannot be suppressed (see
https://github.com/jqno/equalsverifier/pull/627). In practice,
`hashCode` should not matter, so Map-conformant implementation is
acceptable.

`KeyedWindowResult.hashCode()` should never be used to determine
partition id, so the change should not affect any old snapshots.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/6980
Supersedes: https://github.com/hazelcast/hazelcast-mono/pull/382
GitOrigin-RevId: 6b047f3fd1b3906f33947d4ea09a5843c6745f50