Add getCPGroupIds and getCPObjectInfos to CPSubsystem [HZ-3937] [DMT] (#743)
This exposes already existing API
`CPSubsystemManagementService#getCPGroupIds` in CPSubsystem, where it is
easier to use and is supported both by the member and client.

This API provides access to all active group ids, which can then be used
to list all data structures within the group. This API will be sent in a
separate PR.

Original Protocol API:
https://github.com/hazelcast/hazelcast-client-protocol/pull/504

Master PRs:
https://github.com/hazelcast/hazelcast-mono/pull/591
https://github.com/hazelcast/hazelcast-mono/pull/615

GitOrigin-RevId: cd5b9d7d461ecac5adef47d645aeca5fd8c4091f