Fix JobStatusListener registration race [HZ-4330] [HZ-2573] (#673)
When execution of registration/deregistration operation for
`JobStatusListener` was offloaded to event striped executor, we did not
wait for the operation completion (offload is asynchronous). This caused
race conditions in tests but might affect users as well: job listener
existed when it was not expected to exist. In particular, after
registration operation returned, the listener might still not exist for
a while and subsequent deregistration might fail to remove it.

Fixes HZ-4330
Fixes HZ-2573
Fixes hazelcast/hazelcast#24665
Fixes hazelcast/hazelcast-enterprise#6634

GitOrigin-RevId: 053e311abe5f24a177a96502819035ad0de1606b