Improve flakey instance selection logic in `com.hazelcast.topic.impl.reliable.LossToleranceTest` [HZ-4346] (#716)
Summary - avoid scenario where `topicBackupInstance` could be `null`,
causing a [a test
failure](https://github.com/hazelcast/hazelcast/issues/26228).

Detail:
`com.hazelcast.topic.impl.reliable.LossToleranceTest` starts two
instances, and checks each of them to work out which is the owning
instance (and by implication, the backup instance).

This check is not atomic and given the right conditions, it's
__possible__ that both instances could be the owners at
 different times.

This was causing a `null` `topicBackupInstance`, leading to [a test
failure](https://github.com/hazelcast/hazelcast/issues/26228).

I believe those conditions could be summarized as:
1. Owning instance fails to respond to communication
2. Eventually this causes it to be removed from the cluster
3. This causes the backup instance to be promoted to the owner

But I am unsure as to the root cause of that scenario.

In any case, this pattern is not typically how test code is written -
instead a distributed object is generated explicitly targeted at a given
instance using `randomNameOwnedBy`, which sidesteps the problem of
checking to see which one is the owner - this is at least atomic.

Should the previous conditions occur again, the instances may flip - but
this scenario could occur in any other test and therefore must be of no
consequence.

Also tidied the documentation on `com.hazelcast.partition.Partition`.

Fixes: [HZ-4346](https://hazelcast.atlassian.net/browse/HZ-4346),
https://github.com/hazelcast/hazelcast/issues/26228

[HZ-4346]:
https://hazelcast.atlassian.net/browse/HZ-4346?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: da2a0eed1d394739c12a0e2dbc1d3a54d171c5ac