Skip WAN republishing checks when republishing is not used [HZ-4343] (#728)
As identified in HZ-4343, it is currently not possible to WAN replicate
from 1 cluster to multiple other clusters when all clusters share the
same cluster name. This is because the
`AbstractWanPublisher#publishReplicationEventInternal` method checks and
maintains each batch event's list of cluster names, returning early if
the list already contains the target cluster's name.

This check is done to prevent infinite loops when forwarding
(republishing) is enabled - however, when republishing is not enabled,
it is not possible to create an infinite loop scenario as the target
does not try to produce additional WAN events.

This commit simply introduces a new method to `WanPublisher` which is
specific for republished replication, and is called from the
`WanPutOperation` which handles forwarded WAN events. This allows us to
skip these cluster name checks if the event is not being republished.

This commit also adds to the javadoc of `WanReplicationRef` to show the
default value of republishing WAN events, and a simple regression test to
validate functionality.

Fixes https://hazelcast.atlassian.net/browse/HZ-4343

GitOrigin-RevId: 4761a525314f9dde0324de647eea2f5e1359a9cf