Remove partitionId parameter from remoteReplicatedMapEntryViews (#799)
Usually the sources don't have partitions in their API, but handle the
partition assignment transparently for the user.

Also the stage-per-partition in replicated map job led to high number of
source processors holding a reference to the client, which were slow to
be released. This lead to several test failures reporting that clients
weren't closed at the end of the test (see
http://jenkins.hazelcast.com/job/Hazelcast-mono-EE-pr-builder-data-migration/101/testReport/)

GitOrigin-RevId: 0f59263b08ebed487f99c40ffc3326332b2152c4