Move Boolean type to slow SQL index test suite [HZ-1880] (#798)
In this patch we move `BOOLEAN` type to slow SQL index test suite to
decrease PR builder time and decrease a probability to receive a timeout
for PR builder.
Overall, `BOOLEAN` is used way rarely for indexing, unlike `INT` and
`VARCHAR`, and it is not critical for us to move it to slow test suite.

GitOrigin-RevId: 714a35fada499ca2a73919040c1fdc1c39965de2