Mock interface instead of class in MapWanSyncRESTTest [HZ-1993] (#845)
Fixes https://hazelcast.atlassian.net/browse/HZ-1993

The problem is that Mockito occasionally fails to mock the class and
instead calls the real method implementation.
To address this, let's try to mock the interface instead of the class.
By mocking the interface, the issue of invoking real methods is avoided.

To be able to mock the interface the following was done:
was added the used getSyncResult method to the WanReplicationService
interface.
WanSyncStateResult now is an interface located in the os rep.
EnterpriseWanSyncStateResult is an enterprise implementation of
WanSyncStateResult located in ee rep.

So we can mock WanReplicationService and WanSyncStateResult interfaces
instead of EnterpriseWanReplicationService and
EnterpriseWanSyncStateResult.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases
GitOrigin-RevId: 04456a532777580e52dc09b070e596342888fe82