[TPC] Remove hardcoded port number specification from `AsyncServerSocketTest` to improve test stability (#785)
`NioAsyncServerSocketTest.test_createCloseLoop_withSameReactor` is
[regularly failing](https://github.com/hazelcast/hazelcast/issues/24002)
with:
```
java.io.UncheckedIOException: Failed to bind to /127.0.0.1:5000
```

_Some_ of the tests are hardcoded with specific (non-overlapping) ports
- within the scope of this test this _should_ be ok, but there are no
guarantees that these ports will actually be available:
- Another application could've already bound to those ports
- Another test could be using those ports
- If the same tests are being run in parallel that might cause issues

Instead of hardcoding an explicit port, it seems safer to instead use
any free port.

The
[`InetSocketAddress`Javadoc](https://docs.oracle.com/javase/8/docs/api/java/net/InetSocketAddress.html#InetSocketAddress-java.lang.String-int-)
says:
> A valid port value is between 0 and 65535. A port number of zero will
let the system pick up an ephemeral port in a bind operation.

I _believe_ that this performs a similar operation as
[SocketAddress](https://docs.oracle.com/javase/7/docs/api/java/net/ServerSocket.html#ServerSocket%28int%29)
where an unused port is automatically allocated.

I am **unable** to reproduce the test flake before this change to
evaluate if this change addresses it.

Changes:
- updated any hardcoded port numbers to `0`.
- converted the test to JUnit5
- added `assertDoesNotThrow` on some tests to address a Sonar warning
(`S2699 - Tests should include assertions`)

Fixes: https://github.com/hazelcast/hazelcast/issues/24002
GitOrigin-RevId: 756a2edfe5183e39ff3ab5c7a1d60b5ebe860030